Refactor RuleConfiguration to use Spring ConfigurationProperties
